Мелкие правки.

+++ OnlineShop/OnlineShopWebApp/Controllers/CartController.cs
@@ -45,11 +45,8 @@ namespace OnlineShopWebApp.Controllers
 
         public IActionResult Delete(int productId)
         {
-            if (productId > 0)
-            {
-                if (cart != null)
-                    cart = cartStorage.DeleteItem(productId);
-            }
+            if (cart != null && productId > 0)
+                cart = cartStorage.DeleteItem(productId);
 
             return RedirectToAction("Index");
         }
@@ -60,13 +57,13 @@ namespace OnlineShopWebApp.Controllers
             return RedirectToAction("Index");
         }
 
-        public IActionResult IncreaseProductCount(int productId, int quantity = 1)
+        public IActionResult Increase(int productId, int quantity = 1)
         {
             cart = cartStorage.Increase(productId, quantity);
             return RedirectToAction("Index");
         }
 
-        public IActionResult ReduceProductCount(int productId, int quantity = 1)
+        public IActionResult Reduce(int productId, int quantity = 1)
         {
             cart = cartStorage.Reduce(productId, quantity);
             return RedirectToAction("Index");
